Check whether Future was cancelled immediately before setting its result
Python will raise an InvalidStateException if Future.set_value or
Future.set_exception are called even though the Future has already been
cancelled.

This commit moves the check whether the Future has been cancelled yet to
happen immediately and atomically  before calls to Future.set_result and
Future.set_exception. This avoid a "time-of-check to time-of-use" (TOCTOU)
race condition, which can currently occur if the Future is cancelled between
the call to Future.cancelled() and the point in time when the thread-safe
call to Future.set_result or Future.set_exception is actually scheduled to
run.